foxxxx

+++ event/MessageAllow.go
@@ -2,8 +2,8 @@ package event
 
 type MessageAllow struct {
 	Event
-	UserID float64 `json:"user_id" map:"user_id"`
-	Key    string  `json:"key" map:"key"`
+	UserID int    `json:"user_id" map:"user_id"`
+	Key    string `json:"key" map:"key"`
 }
 
 func (m MessageAllow) GetName() string {
