+++ event/MessageDeny.go
@@ -2,7 +2,7 @@ package event
 
 type MessageDeny struct {
 	Event
-	UserID float64 `json:"user_id" map:"user_id"`
+	UserID int `json:"user_id" map:"user_id"`
 }
 
 func (m MessageDeny) GetName() string {
