+++ vk/object/Graffiti.go
@@ -1,7 +1,7 @@
 package object
 
 type Graffiti struct {
-	Source string  `json:"src" map:"src"`
-	Width  float64 `json:"width" map:"width"`
-	Height float64 `json:"height" map:"height"`
+	Source string `json:"src" map:"src"`
+	Width  int    `json:"width" map:"width"`
+	Height int    `json:"height" map:"height"`
 }
