+++ vk/object/Photo.go
@@ -1,21 +1,21 @@
 package object
 
 type PhotoSize struct {
-	Type   string  `json:"type" map:"type"`
-	Url    string  `json:"url" map:"url"`
-	Width  float64 `json:"width" map:"width"`
-	Height float64 `json:"height" map:"height"`
+	Type   string `json:"type" map:"type"`
+	Url    string `json:"url" map:"url"`
+	Width  int    `json:"width" map:"width"`
+	Height int    `json:"height" map:"height"`
 }
 
 type Photo struct {
-	ID      float64     `json:"id" map:"id"`
-	AlbumID float64     `json:"album_id" map:"album_id"`
-	OwnerID float64     `json:"owner_id" map:"owner_id"`
+	ID      int         `json:"id" map:"id"`
+	AlbumID int         `json:"album_id" map:"album_id"`
+	OwnerID int         `json:"owner_id" map:"owner_id"`
 	Text    string      `json:"text" map:"text"`
-	Date    float64     `json:"date" map:"date"`
+	Date    int         `json:"date" map:"date"`
 	Size    []PhotoSize `json:"sizes" map:"sizes"`
-	Width   float64     `json:"width" map:"width"`
-	Height  float64     `json:"height" map:"height"`
+	Width   int         `json:"width" map:"width"`
+	Height  int         `json:"height" map:"height"`
 
 	biggerImageUrl string
 }
@@ -25,7 +25,7 @@ func (p *Photo) GetBiggerImageUrl() (currentUrl string) {
 		return p.biggerImageUrl
 	}
 
-	var maxSize float64 = 0
+	var maxSize int = 0
 	for _, size := range p.Size {
 		crntSize := size.Width * size.Height
 		if crntSize > maxSize {
