+++ vk/object/ResponseError.go
@@ -1,7 +1,7 @@
 package object
 
 type ResponseError struct {
-	ErrorCode float64        `json:"error_code"`
+	ErrorCode int            `json:"error_code"`
 	Error     string         `json:"error_msg"`
 	Params    []RequestParam `json:"request_params"`
 }
